+++ lib/generators/templates/seeds.rb
@@ -1,0 +1,2 @@
+# Load Integral seedings
+Integral::Engine.load_seed