+++ CHANGELOG.md
@@ -18,6 +18,7 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 - Moved list edit form to /edit route & removed show route
 - Dashboard displays most recent post & page
+- Post data appears in dashboard widgets when blog is enabled
 
 ## v1.0.1